+++ src/index.ts
@@ -1,10 +1,4 @@
-export type {
-	CommandLine,
-	CommandLineOutputOptions,
-} from './CommandLine/CommandLine.ts'
-export type { Exception } from './Exception/Exception.ts'
-export type { FileSystem } from './FileSystem/FileSystem.ts'
-export type { StackTrace } from './StackTrace/StackTrace.ts'
-export type { Streams as CommandLineStreams } from './CommandLine/Streams.ts'
+export type { CommandLine } from './CommandLine/CommandLine.ts'
+export type { VirtualFileSystem } from './FileSystem/VirtualFileSystem.ts'
 export { LocalCommandLine } from './CommandLine/LocalCommandLine.ts'
 export { LocalFileSystem } from './FileSystem/LocalFileSystem.ts'