+++ CHANGELOG.md
@@ -18,9 +18,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+-   Rename `FileSystem` interface to `VirtualFileSystem` to avoid conflict with File System API.
+
 ### Deprecated
 
 ### Removed
+
+-   Do not export `Exception` utility class since it is for internal use.
+-   Remove `StackTrace` utility class since it is no longer needed.
 
 ### Fixed